refactor(server): extract application tracker routes (7/8 server split)
Moves the 6 /api/applications/* endpoints to their own Blueprint:
- GET    /api/applications                  (with optional ?status= filter)
- POST   /api/applications                  (upsert by external_id)
- DELETE /api/applications/<ext_id>         (soft-delete: status='removed')
- PATCH  /api/applications/<ext_id>         (partial update)
- GET    /api/applications/company/<name>   (company history, case-insensitive)
- GET    /api/applications/export           (JSON download for backup)

Uses the same live-DB_PATH lookup idiom as profile_routes:
  from core import config as _config
  # later: get_conn(_config.DB_PATH)

test_applications_api.py works unchanged (9 tests pass) — its fixture
must already monkeypatch the right module attribute via setenv before
the route imports happen. Good design call by whoever wrote those tests.

server.py: 577 → 382 LOC (-195). Largest single extraction in the split.

Tests: 219 passed (unchanged).
